backend/payment-sbp/requirements.txt
psycopg2-binary>=2.9.9
qrcode>=7.4.2
pillow>=10.0.0
